+++ RELEASE_NOTES.md
@@ -1,5 +1,17 @@
 # _Sharly Chess_ release notes
 
+## Events
+
+- Import events from SCE files (3.6.0)
+
+## Tournaments
+
+- Distribute the players among the tournaments (3.6.0)
+
+## Players
+
+- Import players from CSV files (3.6.0)
+
 ## Pairings
 
 - Upgrade bbpPairings engine to v6.0.0 (3.6.0)